Move Canvas and all subclasses to canvas folder
