+++ lib/BMPBitmap.js
@@ -162,7 +162,6 @@ module.exports = class BMPBitmap extends Bitmap {
                     4/*bytes*/) * 4/*bytes*/;
             fileBuffer = new Buffer(numberOfBytesPerLine);
             for (let y = header.height - 1; y >= 0; y--) {
-                //noinspection JSDuplicatedDeclaration
                 if (fs.readSync(file, fileBuffer, 0, numberOfBytesPerLine, null) !== numberOfBytesPerLine) {
                     hasError = true;
                     errorMessage = `Unexpected end of file in "${filePath}"`;
