+++ lib/Bitmap.js
@@ -1,8 +1,8 @@
 "use strict";
 
 const Endianness = require("./Endianness");
-const Canvas = require("./Canvas");
-const GrayscaleCanvas = require("./GrayscaleCanvas");
+const Canvas = require("./canvas/Canvas");
+const GrayscaleCanvas = require("./canvas/GrayscaleCanvas");
 
 function calculateGradientValue(step, numSteps, leftColor, rightColor) {
     if (step === 0) {
