+++ lib/canvas/Canvas.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const Endianness = require("./Endianness");
+const Endianness = require("../Endianness");
 
 module.exports = class Canvas {
     constructor(width, height, bytesPerPixel, endianness) {
