+++ lib/canvas/ConsecutiveStorageCanvas.js
