+++ lib/canvas/GrayscaleCanvas.js
