+++ lib/canvas/MultiValueStorageCanvas.js
