+++ lib/canvas/RGBCanvas.js
