+++ lib/canvas/SeparatedRGBCanvas.js
