+++ lib/canvas/SeparatedStorageCanvas.js
